# public/swagger.json
{
	"components": {
...
		"schemas": {
			"Types.ObjectId": {
				"type": "string"
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Required_any_": {
				"properties": {},
				"type": "object",
				"description": "Make all properties in T required"
			},
...
			"ISession": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string"
					},
					"serviceProviderID": {
						"type": "string"
					},
					"clientsIDs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"online",
							"home",
							"office"
						]
					},
					"location": {
						"type": "string"
					},
					"attachments": {
						"type": "string"
					},
					"requierments": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"ﬁnished"
						]
					},
					"doctorReferral": {
						"type": "string"
					}
				},
				"required": [
					"sessionType",
					"serviceProviderID",
					"clientsIDs",
					"name",
					"details",
					"startDate",
					"duration",
					"serviceType",
					"location",
					"attachments",
					"requierments",
					"ratings",
					"reviews",
					"sessionFee",
					"payments",
					"status",
					"doctorReferral"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_ISession_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"sessionType": {
						"type": "string"
					},
					"serviceProviderID": {
						"type": "string"
					},
					"clientsIDs": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"name": {
						"type": "string"
					},
					"details": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"serviceType": {
						"type": "string",
						"enum": [
							"online",
							"home",
							"office"
						]
					},
					"location": {
						"type": "string"
					},
					"attachments": {
						"type": "string"
					},
					"requierments": {
						"type": "string"
					},
					"ratings": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"reviews": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"sessionFee": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"status": {
						"type": "string",
						"enum": [
							"initiated",
							"agreed",
							"canceled",
							"ﬁnished"
						]
					},
					"doctorReferral": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUsers": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastNAme": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": [
							"M",
							"F"
						]
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSend",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSettings": {
						"type": "string"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					}
				},
				"required": [
					"email",
					"phoneNumber",
					"profilePicture",
					"firstName",
					"lastNAme",
					"gender",
					"DOB",
					"address",
					"verified",
					"status",
					"accountType",
					"lastLoginDate",
					"accountSettings",
					"languages",
					"maritalStatus"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IUsers_": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"email": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					},
					"firstName": {
						"type": "string"
					},
					"lastNAme": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": [
							"M",
							"F"
						]
					},
					"DOB": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"verified": {
						"type": "string",
						"enum": [
							"notSend",
							"pending",
							"verified"
						]
					},
					"status": {
						"type": "string",
						"enum": [
							"inActive",
							"active",
							"suspended",
							"lost",
							"deleted"
						]
					},
					"accountType": {
						"type": "string",
						"enum": [
							"PT",
							"EM",
							"PA"
						]
					},
					"lastLoginDate": {
						"type": "string",
						"format": "date-time"
					},
					"accountSettings": {
						"type": "string"
					},
					"languages": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maritalStatus": {
						"type": "string",
						"enum": [
							"married",
							"single",
							"divorced",
							"widow"
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "newproj",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/session/{sessionId}": {
			"get": {
				"operationId": "GetSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ISession"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested session in not found"
					}
				},
				"description": "Get a session details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			},
			"delete": {
				"operationId": "DeleteSession",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/ISession"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested session in not found"
					}
				},
				"description": "Delete a session",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			}
		},
		"/session/create": {
			"post": {
				"operationId": "CreateSession",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ISession"
								},
								"examples": {
									"Example 1": {
										"value": {
											"sessionType": "",
											"serviceProviderID": "",
											"clientsIDs": [],
											"name": "",
											"details": "",
											"startDate": "2022-09-02T21:00:00.000Z",
											"duration": 10,
											"serviceType": "online",
											"location": "",
											"attachments": "",
											"requierments": "",
											"ratings": [
												""
											],
											"reviews": [
												""
											],
											"sessionFee": 150,
											"payments": [
												""
											],
											"status": "agreed",
											"doctorReferral": ""
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a session",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISession"
							}
						}
					}
				}
			}
		},
		"/session/update/{sessionId}": {
			"put": {
				"operationId": "UpdateSession",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ISession"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a session",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ISession_"
							}
						}
					}
				}
			}
		},
...
		"/user/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUsers"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "the requested user in not found"
					}
				},
				"description": "Get a user details",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Document_any.any.any_"
										},
										{
											"$ref": "#/components/schemas/IUsers"
										},
										{
											"$ref": "#/components/schemas/Required_any_"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "the requested user in not found"
					}
				},
				"description": "Delete a user",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6300e18d3bbd975cf6459994"
					}
				]
			}
		},
		"/user/create": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUsers"
								},
								"examples": {
									"Example 1": {
										"value": {
											"email": "",
											"phoneNumber": "",
											"profilePicture": "",
											"firstName": "",
											"lastNAme": "",
											"gender": "F",
											"DOB": "2022-09-02T21:00:00.000Z",
											"address": [
												""
											],
											"verified": "notSend",
											"status": "active",
											"accountType": "PA",
											"lastLoginDate": "2022-09-02T21:00:00.000Z",
											"accountSettings": "",
											"languages": [
												""
											],
											"maritalStatus": "married"
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Create a user",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUsers"
							}
						}
					}
				}
			}
		},
		"/user/update/{userId}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "updated",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUsers"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed"
					}
				},
				"description": "Update a user",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IUsers_"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}